Fixed bugs and improved code quality in the Angular application
Removed the standalone property from the @Component decorator of the AppComponent
Removed RouterOutlet and HeroesComponent from the imports array of the @Component decorator of the AppComponent
Included HeroesComponent in the declarations array of the @NgModule decorator in the main module of the application
Included RouterOutlet in the declarations array of the @NgModule decorator in the main module of the application
Included UppercasePipe in the declarations array of the @NgModule decorator in the main module of the application
Included UppercasePipe in the declarations array of the @NgModule decorator in the main module of the application
Included HeroesComponent in the declarations array of the @NgModule decorator in the main module of the application
Included HeroesComponent in the bootstrap array of the @NgModule decorator in the main module of the application
Imported HeroesComponent in the main module of the application
Imported UppercasePipe in the main module of the application
These changes fix the bugs and improve the code quality of the Angular application. The standalone property is removed from the @Component decorator of the AppComponent, RouterOutlet and HeroesComponent are removed from the imports array of the @Component decorator of the AppComponent, HeroesComponent is included in the declarations array of the @NgModule decorator in the main module of the application, RouterOutlet is included in the declarations array of the @NgModule decorator in the main module of the application, UppercasePipe is included in the declarations array of the @NgModule decorator in the main module of the application, UppercasePipe is included in the declarations array of the @NgModule decorator in the main module of the application, HeroesComponent is included in the declarations array of the @NgModule decorator in the main module of the application, HeroesComponent is included in the bootstrap array of the @NgModule decorator in the main module of the application, HeroesComponent is imported in the main module of the application, and UppercasePipe is imported in the main module of the application.

These changes will make the code more maintainable and easier to understand for other developers.