Update AppComponent and HeroesComponent to use standalone feature; add HeroDetailComponent import to HeroesComponent
The diff includes changes to the AppComponent and HeroesComponent to utilize the standalone feature in Angular. The AppComponent imports the HeroesComponent and includes it in the bootstrap array. The HeroesComponent imports the HeroDetailComponent and includes it in the imports array. The HeroesComponent also includes the HEROES constant import. The AppModule has been updated to have an empty declarations array and bootstrap array. The HeroDetailComponent standalone property is also set to true and includes necessary imports.