Update HeroesComponent: add hero list, use standalone components, remove individual hero view
The CSS for the heroes list has been added. The heroes list is now displayed using \*ngFor. Standalone components NgIf, NgFor, and UpperCasePipe have been imported and used in the template. The individual hero view has been removed. The heroes are now stored in a variable and selectedHero is used to display the details of the selected hero. The onSelect method sets the selected hero.