Add HeroDetailComponent to AppModule declarations and imports
This commit adds the HeroDetailComponent to the declarations and imports arrays in the AppModule of the Tour of Heroes application. However, there seems to be a mistake in the code diff as the HeroDetailComponent is added to the imports array, which is not correct. It should be only in the declarations array.